Implement KvsEngine for KvStore
