+++ rust/projects/networking/Cargo.toml
@@ -13,8 +13,6 @@ serde = { version = "1.0.89", features = ["derive"] }
 serde_json = "1.0.39"
 log = "0.4.6"
 env_logger = "0.6.1"
-parking_lot = "0.7.1"
-lock_api = "0.1.5"
 
 [dev-dependencies]
 tempfile = "3.0.7"
