+++ rust/projects/networking/src/engines/mod.rs
@@ -17,3 +17,5 @@ pub trait KvsEngine {
 }
 
 pub mod kvs;
+
+pub use kvs::KvStore;
